Move helper classes to the 'Common' folder.
