+++ src/ResourceManager/Compute/Commands.Compute/Common/HashTableExtensions.cs
@@ -16,9 +16,9 @@ using Microsoft.Azure.Commands.Tags.Model;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Microsoft.Azure.Commands.Compute.Models
+namespace Microsoft.Azure.Commands.Compute.Common
 {
-    public static class HashTableExtension
+    public static class HashTableExtensions
     {
         public static Dictionary<string, string> ToDictionary(this Hashtable[] tags)
         {
